Recreate swapchain when window is resized on dawn_vulkan backend
This patch adds the recreation of the swap chain when we detect the
window is resized on dawn_vulkan backend to workaround a crash issue
when the window is resized but we don't recreate the swap chain.

The swap chain recreation will be enabled on all backends when the
code is ready in Dawn.

Note that Aquarium will crash on Linux when using dawn_vulkan
backend, which needs further investigation.