Update shaderc to latest commit
Extra semicolon warning error is fixed in latest shaderc, and it's a
third party of dawn. So we have to move shaderc to upper folder, and
maintain shaderc by Aquarium.

This patch also add .gitattributes file to specify format of files because
the ending of line should be specified to LF, otherwise it will become
CRLF on Windows.