Stabilize FPS
Because type of time was float which is not precise enough,
elapse was 0 between 2 frames when rendering small amount of
fishes, thus, the value was wrong. So in this patch, float is
replaced by double to calculate fps. In addition, fps is
calculated by average of every 50ms to eliminate divisor is
almost 0.