Revise update and draw logic of Renderpass and non-Renderpass
Renderpass is a feature of advanced APIs like Dawn and D3D12.
In the render loop, the common logic of render pass updates the
uniform buffers firstly, then begin render pass and record draw
cmds and end render pass. The logic is different from OpenGL
style apps in which update and draw are interspersed.