D3D12 renderpass workspace (#57)
* Revise update and draw logic of Renderpass and non-Renderpass

Renderpass is a feature of advanced APIs like Dawn and D3D12.
In the render loop, the common logic of render pass updates the
uniform buffers firstly, then begin render pass and record draw
cmds and end render pass. The logic is different from OpenGL
style apps in which update and draw are interspersed.

* Enable render pass toggle for D3D12
1.Add toggle DisableD3D12RenderPass to D3D12 backend
2.Check if the feature is available
3.Refactor code for non render pass

* Implement D3D12 render pass

1. Show render pass state on imgui.
2. Implement render pass with both MSAA and non-MSAA paths.
3. Implement MSAA for imgui D3D12. The pso of imgui should use
the same MSAA state of the psos of Aquarium models. This also
fixed a bug for non-render-pass MSAA path of D3D12 backend.
Also refactor imgui dawn to match the api of imgui d3d12.
4. Split update and draw part of imgui to two part because
UAV or bindings should not be revised in render pass.

* Fix some bugs
1.Init current fish instance and pre fish instance
2.Remove unnecessary ref of Comptr when calling functions
3.Use command list of contextD3D12 directly instead of auto &
4.Ref of local variable is removed because of the life cycle.