Implement buffer mapping async
Implement automatic buffer allocation logic in buffer manager.
Two queues are maintained, one is MappedBufferList and another is
EnqueuedBufferList.
Repeat 1 to 4 in every frame
Repeat 1 and 2 whenever buffer is needed
1. Whenever a new buffer is required, new buffer is acquired from MappedBufferList
so that the buffer can be reused as much as possible. If a valid buffer is found,
pop the buffer from MappedBufferList and push into EnqueuedBufferList. If no valid
buffer is found in the list, the app will allocate a new mapped buffer and then
push the buffer into EnqueuedBufferList.
2. Then copy data to the buffer pointer, and copy from the staging buffer to
device buffer. The copy commands are record in copy command encoder which is
independent of other rendering commands.
3. After that, ummap the buffers in EnqueuedBufferList and submit the copy commands.
4. In order to avoid blocking main thread, map the buffers by async callback
function and push them into MappedBufferList after success status in the call back
function.