Implement alpha blending toggle for Aquarium (#71)
1. Add alpha blending toggle to Aquarium. By default, alpha
blending is disabled. To enable alpha blending, you can
pass "--enable-alpha-blending=[0, 1]" by cmd args.
2. Replace alpha value of shaders by regex to specify the alpha
value.
3. Imgui uses stb truetype to render characters to gray bitmaps
with format alpha8. The alpha value of the bitmap is 255 or 0.
When alpha blending is enabled, the pixels of the glyphs shows
only if alpha is 255, while the other pixels are rendered as
background color. When alpha blending is disabled, the pixels
are set to white if alpha is 255, and black if alpha is 0.