Add initial CI scripts
These scripts are intended to serve the GitHub Actions workflow
triggered by pull request events. The workflow configuration file will
be added to the master branch in a separate commit soon.

Currently, commits in a pull request are squashed and applied onto
master. In the future, the scripts may be extended to use rebase
instead, as long as each message title matches a certain pattern. In
that case, the commits are going to be checked one by one.

Another potential next step is to handle the repository dispatch
event. This could be useful in firing up a build with customized
configuration.

Some of the scripts can run locally without the GitHub runner
environment, see the comments at the beginning of these files.

See also https://help.github.com/en/actions for more info about GitHub
Actions.