Remove component type from the test matrix
Aquarium is not a huge project with gigabytes of symbols, so component
build brings little benefit. IOW, it's unlikely that is_component_build
needs to be tuned in practice. Since is_component_build defaults to the
value of is_debug, the bot can still provide partial coverage with the
new test matrix.