Massive clang-formatting
Files are considered as originating from 3rd-parties, iff any of the
following is true:

  * the file resides in src/include/d3d12
  * the file is named in the form of lower_case.ext

They are exempted here to make the life of integrating upstream updates
easier in the future.

For first party code, use of clang-format is recommended, but not
enforced. It's almost certain that clang-format may result in visually
worse typography in some places, however that's rare and in general the
code looks better than before. To accomplish this:

  find src \
    -wholename src/include/d3d -prune -o \
    -regex '.+/[a-z0-9_]+\.[a-z0-9]+' -prune -o \
    '(' -name '*.h' -o -name '*.cpp' ')' -exec clang-format -i '{}' ';'