Create backend contexts via helpers
This is a preparation for deprecating the usage of dawn_bindings. To
achieve that, we will add one .cpp file for the dawn_vulkan and
dawn_d3d12 backend, and one .mm file for the dawn_metal backend. Each
backend will be implemented as a subclass of ContextDawn then and
instantiated with the factory pattern.