Temporarily revert ba8af515dd61137893d78c445ef43e5c5405c78c
The plan is to have revisions of 3rd-parties always derived from
Chromium, so they are known to work nicely together. Considering Abseil
becomes a hard dependency of GoogleTest in latest Chromium, which
supersedes cxxopts, relying on both sounds awkward then. But to avoid
doing too much in one step, it's highly desired to separate the cxxopts
removal from the other overhaul. As a result, a carefully chosen range
of Chromium revisions comes out, and that in turn requires the roll-back
here.